Remove redundancies.
We have a dependency on Macros.h anyway, so
make use of it in Nagler.
