+++ src/Tests/InstrEngineTests/NaglerInstrumentationMethod/assertions.h
@@ -1,9 +1,5 @@
 #pragma once
 
-#ifndef IfFailRet
-#define IfFailRet(EXPR) \
-do { if (FAILED(hr = (EXPR))) { AssertFailed("IfFailRet(" #EXPR ") failed"); return hr; } } while (false)
-#endif
 
 // Note, we use const char* because it is easier for cross-plat
 inline void AssertFailed(_In_ const char* message)
